fix hook to cover all use cases not deletion

+++ internal/ent/hooks/memberhelpers.go
@@ -83,9 +83,11 @@ func HookMembershipSelf(table string) ent.Hook {
 				return nil, err
 			}
 
-			if err := deleteOrgMembershipFGATuples(ctx, orgMembership, mutationMember.Client()); err != nil {
-				zerolog.Ctx(ctx).Error().Err(err).Msg("failed to delete FGA tuples after successful DB deletion")
-				return nil, err
+			if table == "org_memberships" && (m.Op().Is(ent.OpDelete | ent.OpDeleteOne)) {
+				if err := deleteOrgMembershipFGATuples(ctx, orgMembership, mutationMember.Client()); err != nil {
+					zerolog.Ctx(ctx).Error().Err(err).Msg("failed to delete FGA tuples after successful DB deletion")
+					return nil, err
+				}
 			}
 
 			return retVal, err
